creation des structures en cours

+++ libft/srcs/size_len_maths/ft_negative_power.c
@@ -6,15 +6,15 @@
 /*   By: nfauconn <nfauconn@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/23 15:42:03 by user42            #+#    #+#             */
-/*   Updated: 2022/11/03 14:44:40 by nfauconn         ###   ########.fr       */
+/*   Updated: 2022/11/03 16:01:23 by nfauconn         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-double	ft_negative_power(int nb, int power)
+float	ft_negative_power(int nb, int power)
 {
-	double	res;
+	float	res;
 
 	power *= -1;
 	res = 1 / ft_power(nb, power);
